requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
Faker==18.6.2
pytest==6.2.5
pytest-xdist==3.2.1
selenium==4.1.0
webdriver-manager==3.8.6
python-dotenv==1.0.0
